fix non-root deployments
If the app is built for a non-root deployment (e.g. by adding `"autoBasePath": true` into `polymer.json`), polymer build updates the `<base>` tag in `index.html` to the relative deployment path (e.g. to `/es5-bundled/`). Since the base tag has effect only on relative URLs, all URLs in the app should be relative in order to resolve correctly in non-root deployments. Absolute URLs ignore the base tag and would result in incorrect URLs (404).

In addition, this commit removes the need to manually set the global `Polymer.rootPath` property. Instead, the property is automatically initialized with the same value that is set into the `<base>` tag during the build.